% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longke-package.R
\docType{package}
\name{longke-package}
\alias{longke-package}
\title{Nonparametric predictive model for sparse and irregular longitudinal data}
\description{
This package contains the function for fitting nonparametric predictive model for sparse and
irregular longitudinal data
}
\references{
Wang S, Kim S, Cho H, Chang W.
Nonparametric predictive model for sparse and irregular longitudinal data. (2023+)
}
